move function to viewicons
