--- src/plugins/saveRoleIcon/README.md
@@ -1,2 +1,0 @@
-# Save Role Icon
-Allows you to save role icons by right-clicking them.
